+++ FA2sp/FA2sp.h
@@ -58,6 +58,8 @@ public:
     static bool FastResize;
     static int RecentFileLimit;
     static int MultiSelectionColor;
+    static bool RandomTerrainObjects;
+    static int MaxVoxelFacing;
 };
 
 class Variables