+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # FINALALERT2 - SP  CHANGELOG
 
+## RELEASE 1.4.2 (2022-10-27)
+- Split `CncVxlRenderText`
+- New ***ExtConfig*** : `RandomTerrainObjects` = **BOOLEAN**, enable it so `random trees` will show all `TerrainTypes`.
+- Fixed a bug when saving `[PreviewPack]` and `[Preview]`
+
 ## RELEASE 1.4.1 (2022-06-18)
 - *This is a **cumulative** update*
 - Reimplemented CTeamTypes message handler